+++ solid/system/cstring.hpp
@@ -21,11 +21,11 @@ struct cstring {
     //! Equivalent to strcmp
     static int cmp(const char* _s1, const char* _s2);
     //! Equivalent to strncmp
-    static int ncmp(const char* _s1, const char* _s2, uint _len);
+    static int ncmp(const char* _s1, const char* _s2, size_t _len);
     //! Equivalent to strcasecmp
     static int casecmp(const char* _s1, const char* _s2);
     //! Equivalent to strncasecmp
-    static int ncasecmp(const char* _s1, const char* _s2, uint _len);
+    static int ncasecmp(const char* _s1, const char* _s2, size_t _len);
     static size_t nlen(const char* s, size_t maxlen);
     template <class T, T C1>
     static T* find(T* _pc)
